Add GetAllDecayScript
Iterates through an array of .wav audio files and calculate the decay time (time between the initial largest peak, and the lowest peak before the baseline is reached), and we plot these along with position data extracted from a CSV file to plot the decay according to a color map on a 3D scatter graph. We do this by getting the low pass filter envelope of the impulse response. This is then a subplot in a time-domain plot. along with a subplot of the array of low pass filter envelopes color mapped according to the distance from the audio source to the source emission position.